Move JS->Amfora API from display/handlers_js.go to amfora_js.go
