+++ display/handlers_x.go
@@ -1,3 +1,4 @@
+//go:build !js
 // +build !js
 
 package display
@@ -5,4 +6,3 @@ package display
 import "github.com/makeworld-the-better-one/amfora/structs"
 
 func observePage(p *structs.Page) {}
-
